Add fallback for whenever system freetype2 is not installed
Rebased.

By default, icedtea uses FT2_CFLAGS and FT2_LIBS, but unlike other commits for
using the system option, this one doesn't have a way of defaulting out if the user
does not have/need system freetype2 installed. That being said, it also does
not automatically export freetype libs and cflags if freetype2 is available on
system. Fix that.